refact: remove unnecessary routes and refact views/* (#9)
