+++ src/utils/bookSourceEditConfig.js
@@ -1,564 +1,588 @@
 export default {
-  base: [
-    {
-      title: "源域名",
-      rows: 1,
-      id: "bookSourceUrl",
-      type: "String",
-      hint: "<必填>通常填写网站主页,例: https://www.qidian.com",
-    },
-    {
-      title: "源类型",
-      rows: 1,
-      id: "bookSourceType",
-      type: "Number",
-      hint: "<必填>0:文本 1:音频 2:图片 3:文件(只提供下载的网站)",
-    },
-    {
-      title: "源名称",
-      rows: 1,
-      id: "bookSourceName",
-      type: "String",
-      hint: "<必填>会显示在源列表",
-    },
-    {
-      title: "源分组",
-      rows: 1,
-      id: "bookSourceGroup",
-      type: "String",
-      hint: "<选填>描述源的特征信息",
-    },
-    {
-      title: "源注释",
-      rows: 1,
-      id: "bookSourceComment",
-      type: "String",
-      hint: "<选填>描述源作者和状态",
-    },
-    {
-      title: "书源变量",
-      rows: 1,
-      id: "variableComment",
-      type: "String",
-      hint: "<选填>书源变量说明",
-    },
-    {
-      title: "登录地址",
-      rows: 1,
-      id: "loginUrl",
-      type: "String",
-      hint: "<选填>填写网站登录网址,仅在需要登录的源有用",
-    },
-    {
-      title: "登录界面",
-      rows: 3,
-      id: "loginUi",
-      type: "String",
-      hint: "<选填>自定义登录界面",
-    },
-    {
-      title: "登录检测",
-      rows: 3,
-      id: "loginCheckJs",
-      type: "String",
-      hint: "<选填>登录检测js",
-    },
-    {
-      title: "封面解密",
-      rows: 3,
-      id: "coverDecodeJs",
-      type: "String",
-      hint: "<选填>封面解密js",
-    },
-    {
-      title: "并发率",
-      rows: 1,
-      id: "concurrentRate",
-      type: "String",
-      hint: "<选填>并发率",
-    },
-    {
-      title: "请求头",
-      rows: 3,
-      id: "header",
-      type: "String",
-      hint: "<选填>客户端标识",
-    },
-    {
-      title: "链接验证",
-      rows: 1,
-      id: "bookUrlPattern",
-      type: "String",
-      hint: "<选填>当详情页URL与源URL的域名不一致时有效，用于添加网址",
-    },
-  ],
-  search: [
-    {
-      title: "搜索地址",
-      rows: 1,
-      id: "searchUrl",
-      type: "String",
-      hint: "[域名可省略]/search.php@kw={{key}}",
-    },
-    {
-      title: "校验文字",
-      rows: 1,
-      id: "ruleSearch_checkKeyWord",
-      type: "String",
-      hint: "校验关键字",
-    },
-    {
-      title: "列表规则",
-      rows: 1,
-      id: "ruleSearch_bookList",
-      type: "String",
-      hint: "选择书籍节点 (规则结果为List<Element>)",
-    },
-    {
-      title: "书名规则",
-      rows: 1,
-      id: "ruleSearch_name",
-      type: "String",
-      hint: "选择节点书名 (规则结果为String)",
-    },
-    {
-      title: "作者规则",
-      rows: 1,
-      id: "ruleSearch_author",
-      type: "String",
-      hint: "选择节点作者 (规则结果为String)",
-    },
-    {
-      title: "分类规则",
-      rows: 1,
-      id: "ruleSearch_kind",
-      type: "String",
-      hint: "选择节点分类信息 (规则结果为String)",
-    },
-    {
-      title: "字数规则",
-      rows: 1,
-      id: "ruleSearch_wordCount",
-      type: "String",
-      hint: "选择节点字数信息 (规则结果为String)",
-    },
-    {
-      title: "最新章节",
-      rows: 1,
-      id: "ruleSearch_lastChapter",
-      type: "String",
-      hint: "选择节点最新章节 (规则结果为String)",
-    },
-    {
-      title: "简介规则",
-      rows: 1,
-      id: "ruleSearch_intro",
-      type: "String",
-      hint: "选择节点书籍简介 (规则结果为String)",
-    },
-    {
-      title: "封面规则",
-      rows: 1,
-      id: "ruleSearch_coverUrl",
-      type: "String",
-      hint: "选择节点书籍封面 (规则结果为String类型的url)",
-    },
-    {
-      title: "详情地址",
-      rows: 1,
-      id: "ruleSearch_bookUrl",
-      type: "String",
-      hint: "选择书籍详情页网址 (规则结果为String类型的url)",
-    },
-  ],
-  find: [
-    {
-      title: "发现地址",
-      rows: 6,
-      id: "exploreUrl",
-      type: "String",
-      hint: "内容能显示在发现菜单\n每行一条发现分类(网址域名可省略)，例：\n名称1::网址(Url)1\n名称2::网址(Url)2\n...",
-    },
-    {
-      title: "列表规则",
-      rows: 1,
-      id: "ruleExplore_bookList",
-      type: "String",
-      hint: "选择书籍节点 (规则结果为List<Element>)",
-    },
-    {
-      title: "书名规则",
-      rows: 1,
-      id: "ruleExplore_name",
-      type: "String",
-      hint: "选择节点书名 (规则结果为String)",
-    },
-    {
-      title: "作者规则",
-      rows: 1,
-      id: "ruleExplore_author",
-      type: "String",
-      hint: "选择节点作者 (规则结果为String)",
-    },
-    {
-      title: "分类规则",
-      rows: 1,
-      id: "ruleExplore_kind",
-      type: "String",
-      hint: "选择节点分类信息 (规则结果为String)",
-    },
-    {
-      title: "字数规则",
-      rows: 1,
-      id: "ruleExplore_wordCount",
-      type: "String",
-      hint: "选择节点字数信息 (规则结果为String)",
-    },
-    {
-      title: "最新章节",
-      rows: 1,
-      id: "ruleExplore_lastChapter",
-      type: "String",
-      hint: "选择节点最新章节 (规则结果为String)",
-    },
-    {
-      title: "简介规则",
-      rows: 1,
-      id: "ruleExplore_intro",
-      type: "String",
-      hint: "选择节点书籍简介 (规则结果为String)",
-    },
-    {
-      title: "封面规则",
-      rows: 1,
-      id: "ruleExplore_coverUrl",
-      type: "String",
-      hint: "选择节点书籍封面 (规则结果为String类型的url)",
-    },
-    {
-      title: "详情地址",
-      rows: 1,
-      id: "ruleExplore_bookUrl",
-      type: "String",
-      hint: "选择书籍详情页网址 (规则结果为String类型的url)",
-    },
-  ],
-  detail: [
-    {
-      title: "预处理",
-      rows: 3,
-      id: "ruleBookInfo_init",
-      type: "String",
-      hint: "用于加速详情信息检索，只支持AllInOne规则",
-    },
-    {
-      title: "书名规则",
-      rows: 1,
-      id: "ruleBookInfo_name",
-      type: "String",
-      hint: "选择节点书名 (规则结果为String)",
-    },
-    {
-      title: "作者规则",
-      rows: 1,
-      id: "ruleBookInfo_author",
-      type: "String",
-      hint: "选择节点作者 (规则结果为String)",
-    },
-    {
-      title: "分类规则",
-      rows: 1,
-      id: "ruleBookInfo_kind",
-      type: "String",
-      hint: "选择节点分类信息 (规则结果为String)",
-    },
-    {
-      title: "字数规则",
-      rows: 1,
-      id: "ruleBookInfo_wordCount",
-      type: "String",
-      hint: "选择节点字数信息 (规则结果为String)",
-    },
-    {
-      title: "最新章节",
-      rows: 1,
-      id: "ruleBookInfo_lastChapter",
-      type: "String",
-      hint: "选择节点最新章节 (规则结果为String)",
-    },
-    {
-      title: "简介规则",
-      rows: 1,
-      id: "ruleBookInfo_intro",
-      type: "String",
-      hint: "选择节点书籍简介 (规则结果为String)",
-    },
-    {
-      title: "封面规则",
-      rows: 1,
-      id: "ruleBookInfo_coverUrl",
-      type: "String",
-      hint: "选择节点书籍封面 (规则结果为String类型的url)",
-    },
-    {
-      title: "目录地址",
-      rows: 1,
-      id: "ruleBookInfo_tocUrl",
-      type: "String",
-      hint: "选择书籍详情页网址 (规则结果为String类型的url, 与详情页相同时可省略)",
-    },
-    {
-      title: "下载URL",
-      rows: 1,
-      id: "ruleBookInfo_downloadUrls",
-      type: "String",
-      hint: "文件类书源下载地址 (规则结果为String类型的url, 多个链接返回数组)",
-    },
-    {
-      title: "修改书籍",
-      rows: 1,
-      id: "ruleBookInfo_canReName",
-      type: "String",
-      hint: "允许修改书名作者(规则结果为String类型, 默认不允许)",
-    },
-  ],
-  directory: [
-    {
-      title: "更新前JS",
-      rows: 3,
-      id: "ruleToc_preUpdateJs",
-      type: "String",
-      hint: "更新前JS 动态更新书籍目录链接",
-    },
-    {
-      title: "列表规则",
-      rows: 3,
-      id: "ruleToc_chapterList",
-      type: "String",
-      hint: "选择目录列表的章节节点 (规则结果为List<Element>)",
-    },
-    {
-      title: "章节名称",
-      rows: 1,
-      id: "ruleToc_chapterName",
-      type: "String",
-      hint: "选择章节名称 (规则结果为String)",
-    },
-    {
-      title: "章节地址",
-      rows: 1,
-      id: "ruleToc_chapterUrl",
-      type: "String",
-      hint: "选择章节链接 (规则结果为String类型的Url)",
-    },
-    {
-      title: "卷名标识",
-      rows: 1,
-      id: "ruleToc_isVolume",
-      type: "String",
-      hint: "章节名称是否是卷名 (规则结果为Bool)",
-    },
-    {
-      title: "收费标识",
-      rows: 1,
-      id: "ruleToc_isVip",
-      type: "String",
-      hint: "章节是否为VIP章节 (规则结果为Bool)",
-    },
-    {
-      title: "购买标识",
-      rows: 1,
-      id: "ruleToc_isPay",
-      type: "String",
-      hint: "章节是否为已购买 (规则结果为Bool)",
-    },
-    {
-      title: "章节信息",
-      rows: 1,
-      id: "ruleToc_updateTime",
-      type: "String",
-      hint: "选择章节信息 (规则结果为String)",
-    },
-    {
-      title: "翻页规则",
-      rows: 1,
-      id: "ruleToc_nextTocUrl",
-      type: "String",
-      hint: "选择目录下一页链接 (规则结果为List<String>或String)",
-    },
-  ],
-  content: [
-    {
-      title: "脚本注入",
-      rows: 3,
-      id: "ruleContent_webJs",
-      type: "String",
-      hint: "注入javascript，用于模拟鼠标点击等，必须有返回值，一般为String类型",
-    },
-    {
-      title: "正文规则",
-      rows: 1,
-      id: "ruleContent_content",
-      type: "String",
-      hint: "选择正文内容 (规则结果为String)",
-    },
-    {
-      title: "翻页规则",
-      rows: 1,
-      id: "ruleContent_nextContentUrl",
-      type: "String",
-      hint: "选择下一分页(不是下一章)链接 (规则结果为String类型的Url)",
-    },
-    {
-      title: "资源正则",
-      rows: 1,
-      id: "ruleContent_sourceRegex",
-      type: "String",
-      hint: "匹配资源的url特征，用于嗅探",
-    },
-    {
-      title: "替换规则",
-      rows: 1,
-      id: "ruleContent_replaceRegex",
-      type: "String",
-      hint: "多页内容合并后替换，用于正文净化",
-    },
-    {
-      title: "图片样式",
-      rows: 1,
-      id: "ruleContent_imageStyle",
-      type: "String",
-      hint: "FULL:铺满 不填:默认样式",
-    },
-    {
-      title: "购买操作",
-      rows: 1,
-      id: "ruleContent_payAction",
-      type: "String",
-      hint: "填写JavaScript 返回购买链接或者调用购买接口",
-    },
-    {
-      title: "图片解密",
-      rows: 1,
-      id: "ruleContent_imageDecode",
-      type: "String",
-      hint: "填写JavaScript 返回解密图片的bytes ",
-    },
-  ],
-  review: [
-    {
-      title: "段评URL",
-      rows: 1,
-      id: "ruleReview_reviewUrl",
-      type: "String",
-      hint: "段评URL",
-    },
-    {
-      title: "发布者头像",
-      rows: 1,
-      id: "ruleReview_avatarRule",
-      type: "String",
-      hint: "段评发布者头像",
-    },
-    {
-      title: "段评内容",
-      rows: 1,
-      id: "ruleReview_contentRule",
-      type: "String",
-      hint: "段评内容",
-    },
-    {
-      title: "发布时间",
-      rows: 1,
-      id: "ruleReview_postTimeRule",
-      type: "String",
-      hint: "段评发布时间",
-    },
-    {
-      title: "回复URL",
-      rows: 1,
-      id: "ruleReview_reviewQuoteUrl",
-      type: "String",
-      hint: "获取段评回复URL",
-    },
-    {
-      title: "点赞URL",
-      rows: 1,
-      id: "ruleReview_voteUpUrl",
-      type: "String",
-      hint: "点赞URL",
-    },
-    {
-      title: "点踩URL",
-      rows: 1,
-      id: "ruleReview_voteDownUrl",
-      type: "String",
-      hint: "点踩URL",
-    },
-    {
-      title: "发送回复",
-      rows: 1,
-      id: "ruleReview_postReviewUrl",
-      type: "String",
-      hint: "发送回复URL",
-    },
-    {
-      title: "回复段评",
-      rows: 1,
-      id: "ruleReview_postQuoteUrl",
-      type: "String",
-      hint: "发送回复段评URL",
-    },
-    {
-      title: "删除段评",
-      rows: 1,
-      id: "ruleReview_deleteUrl",
-      type: "String",
-      hint: "删除段评URL",
-    },
-  ],
-  other: [
-    {
-      title: "启用搜索",
-      rows: 1,
-      id: "enabled",
-      type: "Boolean",
-      hint: "启用: true  关闭: false (可选,默认true)",
-    },
-    {
-      title: "启用发现",
-      rows: 1,
-      id: "enabledExplore",
-      type: "Boolean",
-      hint: "启用: true  关闭: false (可选,默认true)",
-    },
-    {
-      title: "启用段评",
-      rows: 1,
-      id: "enabledReview",
-      type: "Boolean",
-      hint: "启用: true  关闭: false (可选,默认false)",
-    },
-    {
-      title: "Cookie",
-      rows: 1,
-      id: "enabledCookieJar",
-      type: "Boolean",
-      hint: "保存返回头的Cookie 启用: true  关闭: false (可选,默认false)",
-    },
-    {
-      title: "搜索权重",
-      rows: 1,
-      id: "weight",
-      type: "Number",
-      hint: "整数: 0~N (可选,默认0) | 数字越大越靠前",
-    },
-    {
-      title: "排序编号",
-      rows: 1,
-      id: "customOrder",
-      type: "Number",
-      hint: "整数: 0~N (可选,默认0) | 数字越小越靠前",
-    },
-  ],
+  base: {
+    name: "基础",
+    children: [
+      {
+        title: "源域名",
+        rows: 1,
+        id: "bookSourceUrl",
+        type: "String",
+        hint: "<必填>通常填写网站主页,例: https://www.qidian.com",
+      },
+      {
+        title: "源类型",
+        rows: 1,
+        id: "bookSourceType",
+        type: "Number",
+        hint: "<必填>0:文本 1:音频 2:图片 3:文件(只提供下载的网站)",
+      },
+      {
+        title: "源名称",
+        rows: 1,
+        id: "bookSourceName",
+        type: "String",
+        hint: "<必填>会显示在源列表",
+      },
+      {
+        title: "源分组",
+        rows: 1,
+        id: "bookSourceGroup",
+        type: "String",
+        hint: "<选填>描述源的特征信息",
+      },
+      {
+        title: "源注释",
+        rows: 1,
+        id: "bookSourceComment",
+        type: "String",
+        hint: "<选填>描述源作者和状态",
+      },
+      {
+        title: "书源变量",
+        rows: 1,
+        id: "variableComment",
+        type: "String",
+        hint: "<选填>书源变量说明",
+      },
+      {
+        title: "登录地址",
+        rows: 1,
+        id: "loginUrl",
+        type: "String",
+        hint: "<选填>填写网站登录网址,仅在需要登录的源有用",
+      },
+      {
+        title: "登录界面",
+        rows: 3,
+        id: "loginUi",
+        type: "String",
+        hint: "<选填>自定义登录界面",
+      },
+      {
+        title: "登录检测",
+        rows: 3,
+        id: "loginCheckJs",
+        type: "String",
+        hint: "<选填>登录检测js",
+      },
+      {
+        title: "封面解密",
+        rows: 3,
+        id: "coverDecodeJs",
+        type: "String",
+        hint: "<选填>封面解密js",
+      },
+      {
+        title: "并发率",
+        rows: 1,
+        id: "concurrentRate",
+        type: "String",
+        hint: "<选填>并发率",
+      },
+      {
+        title: "请求头",
+        rows: 3,
+        id: "header",
+        type: "String",
+        hint: "<选填>客户端标识",
+      },
+      {
+        title: "链接验证",
+        rows: 1,
+        id: "bookUrlPattern",
+        type: "String",
+        hint: "<选填>当详情页URL与源URL的域名不一致时有效，用于添加网址",
+      },
+    ],
+  },
+  search: {
+    name: "搜索",
+    children: [
+      {
+        title: "搜索地址",
+        rows: 1,
+        id: "searchUrl",
+        type: "String",
+        hint: "[域名可省略]/search.php@kw={{key}}",
+      },
+      {
+        title: "校验文字",
+        rows: 1,
+        id: "ruleSearch_checkKeyWord",
+        type: "String",
+        hint: "校验关键字",
+      },
+      {
+        title: "列表规则",
+        rows: 1,
+        id: "ruleSearch_bookList",
+        type: "String",
+        hint: "选择书籍节点 (规则结果为List<Element>)",
+      },
+      {
+        title: "书名规则",
+        rows: 1,
+        id: "ruleSearch_name",
+        type: "String",
+        hint: "选择节点书名 (规则结果为String)",
+      },
+      {
+        title: "作者规则",
+        rows: 1,
+        id: "ruleSearch_author",
+        type: "String",
+        hint: "选择节点作者 (规则结果为String)",
+      },
+      {
+        title: "分类规则",
+        rows: 1,
+        id: "ruleSearch_kind",
+        type: "String",
+        hint: "选择节点分类信息 (规则结果为String)",
+      },
+      {
+        title: "字数规则",
+        rows: 1,
+        id: "ruleSearch_wordCount",
+        type: "String",
+        hint: "选择节点字数信息 (规则结果为String)",
+      },
+      {
+        title: "最新章节",
+        rows: 1,
+        id: "ruleSearch_lastChapter",
+        type: "String",
+        hint: "选择节点最新章节 (规则结果为String)",
+      },
+      {
+        title: "简介规则",
+        rows: 1,
+        id: "ruleSearch_intro",
+        type: "String",
+        hint: "选择节点书籍简介 (规则结果为String)",
+      },
+      {
+        title: "封面规则",
+        rows: 1,
+        id: "ruleSearch_coverUrl",
+        type: "String",
+        hint: "选择节点书籍封面 (规则结果为String类型的url)",
+      },
+      {
+        title: "详情地址",
+        rows: 1,
+        id: "ruleSearch_bookUrl",
+        type: "String",
+        hint: "选择书籍详情页网址 (规则结果为String类型的url)",
+      },
+    ],
+  },
+  find: {
+    name: "发现",
+    children: [
+      {
+        title: "发现地址",
+        rows: 6,
+        id: "exploreUrl",
+        type: "String",
+        hint: "内容能显示在发现菜单\n每行一条发现分类(网址域名可省略)，例：\n名称1::网址(Url)1\n名称2::网址(Url)2\n...",
+      },
+      {
+        title: "列表规则",
+        rows: 1,
+        id: "ruleExplore_bookList",
+        type: "String",
+        hint: "选择书籍节点 (规则结果为List<Element>)",
+      },
+      {
+        title: "书名规则",
+        rows: 1,
+        id: "ruleExplore_name",
+        type: "String",
+        hint: "选择节点书名 (规则结果为String)",
+      },
+      {
+        title: "作者规则",
+        rows: 1,
+        id: "ruleExplore_author",
+        type: "String",
+        hint: "选择节点作者 (规则结果为String)",
+      },
+      {
+        title: "分类规则",
+        rows: 1,
+        id: "ruleExplore_kind",
+        type: "String",
+        hint: "选择节点分类信息 (规则结果为String)",
+      },
+      {
+        title: "字数规则",
+        rows: 1,
+        id: "ruleExplore_wordCount",
+        type: "String",
+        hint: "选择节点字数信息 (规则结果为String)",
+      },
+      {
+        title: "最新章节",
+        rows: 1,
+        id: "ruleExplore_lastChapter",
+        type: "String",
+        hint: "选择节点最新章节 (规则结果为String)",
+      },
+      {
+        title: "简介规则",
+        rows: 1,
+        id: "ruleExplore_intro",
+        type: "String",
+        hint: "选择节点书籍简介 (规则结果为String)",
+      },
+      {
+        title: "封面规则",
+        rows: 1,
+        id: "ruleExplore_coverUrl",
+        type: "String",
+        hint: "选择节点书籍封面 (规则结果为String类型的url)",
+      },
+      {
+        title: "详情地址",
+        rows: 1,
+        id: "ruleExplore_bookUrl",
+        type: "String",
+        hint: "选择书籍详情页网址 (规则结果为String类型的url)",
+      },
+    ],
+  },
+  detail: {
+    name: "详情",
+    children: [
+      {
+        title: "预处理",
+        rows: 3,
+        id: "ruleBookInfo_init",
+        type: "String",
+        hint: "用于加速详情信息检索，只支持AllInOne规则",
+      },
+      {
+        title: "书名规则",
+        rows: 1,
+        id: "ruleBookInfo_name",
+        type: "String",
+        hint: "选择节点书名 (规则结果为String)",
+      },
+      {
+        title: "作者规则",
+        rows: 1,
+        id: "ruleBookInfo_author",
+        type: "String",
+        hint: "选择节点作者 (规则结果为String)",
+      },
+      {
+        title: "分类规则",
+        rows: 1,
+        id: "ruleBookInfo_kind",
+        type: "String",
+        hint: "选择节点分类信息 (规则结果为String)",
+      },
+      {
+        title: "字数规则",
+        rows: 1,
+        id: "ruleBookInfo_wordCount",
+        type: "String",
+        hint: "选择节点字数信息 (规则结果为String)",
+      },
+      {
+        title: "最新章节",
+        rows: 1,
+        id: "ruleBookInfo_lastChapter",
+        type: "String",
+        hint: "选择节点最新章节 (规则结果为String)",
+      },
+      {
+        title: "简介规则",
+        rows: 1,
+        id: "ruleBookInfo_intro",
+        type: "String",
+        hint: "选择节点书籍简介 (规则结果为String)",
+      },
+      {
+        title: "封面规则",
+        rows: 1,
+        id: "ruleBookInfo_coverUrl",
+        type: "String",
+        hint: "选择节点书籍封面 (规则结果为String类型的url)",
+      },
+      {
+        title: "目录地址",
+        rows: 1,
+        id: "ruleBookInfo_tocUrl",
+        type: "String",
+        hint: "选择书籍详情页网址 (规则结果为String类型的url, 与详情页相同时可省略)",
+      },
+      {
+        title: "下载URL",
+        rows: 1,
+        id: "ruleBookInfo_downloadUrls",
+        type: "String",
+        hint: "文件类书源下载地址 (规则结果为String类型的url, 多个链接返回数组)",
+      },
+      {
+        title: "修改书籍",
+        rows: 1,
+        id: "ruleBookInfo_canReName",
+        type: "String",
+        hint: "允许修改书名作者(规则结果为String类型, 默认不允许)",
+      },
+    ],
+  },
+  directory: {
+    name: "目录",
+    children: [
+      {
+        title: "更新前JS",
+        rows: 3,
+        id: "ruleToc_preUpdateJs",
+        type: "String",
+        hint: "更新前JS 动态更新书籍目录链接",
+      },
+      {
+        title: "列表规则",
+        rows: 3,
+        id: "ruleToc_chapterList",
+        type: "String",
+        hint: "选择目录列表的章节节点 (规则结果为List<Element>)",
+      },
+      {
+        title: "章节名称",
+        rows: 1,
+        id: "ruleToc_chapterName",
+        type: "String",
+        hint: "选择章节名称 (规则结果为String)",
+      },
+      {
+        title: "章节地址",
+        rows: 1,
+        id: "ruleToc_chapterUrl",
+        type: "String",
+        hint: "选择章节链接 (规则结果为String类型的Url)",
+      },
+      {
+        title: "卷名标识",
+        rows: 1,
+        id: "ruleToc_isVolume",
+        type: "String",
+        hint: "章节名称是否是卷名 (规则结果为Bool)",
+      },
+      {
+        title: "收费标识",
+        rows: 1,
+        id: "ruleToc_isVip",
+        type: "String",
+        hint: "章节是否为VIP章节 (规则结果为Bool)",
+      },
+      {
+        title: "购买标识",
+        rows: 1,
+        id: "ruleToc_isPay",
+        type: "String",
+        hint: "章节是否为已购买 (规则结果为Bool)",
+      },
+      {
+        title: "章节信息",
+        rows: 1,
+        id: "ruleToc_updateTime",
+        type: "String",
+        hint: "选择章节信息 (规则结果为String)",
+      },
+      {
+        title: "翻页规则",
+        rows: 1,
+        id: "ruleToc_nextTocUrl",
+        type: "String",
+        hint: "选择目录下一页链接 (规则结果为List<String>或String)",
+      },
+    ],
+  },
+  content: {
+    name: "正文",
+    children: [
+      {
+        title: "脚本注入",
+        rows: 3,
+        id: "ruleContent_webJs",
+        type: "String",
+        hint: "注入javascript，用于模拟鼠标点击等，必须有返回值，一般为String类型",
+      },
+      {
+        title: "正文规则",
+        rows: 1,
+        id: "ruleContent_content",
+        type: "String",
+        hint: "选择正文内容 (规则结果为String)",
+      },
+      {
+        title: "翻页规则",
+        rows: 1,
+        id: "ruleContent_nextContentUrl",
+        type: "String",
+        hint: "选择下一分页(不是下一章)链接 (规则结果为String类型的Url)",
+      },
+      {
+        title: "资源正则",
+        rows: 1,
+        id: "ruleContent_sourceRegex",
+        type: "String",
+        hint: "匹配资源的url特征，用于嗅探",
+      },
+      {
+        title: "替换规则",
+        rows: 1,
+        id: "ruleContent_replaceRegex",
+        type: "String",
+        hint: "多页内容合并后替换，用于正文净化",
+      },
+      {
+        title: "图片样式",
+        rows: 1,
+        id: "ruleContent_imageStyle",
+        type: "String",
+        hint: "FULL:铺满 不填:默认样式",
+      },
+      {
+        title: "购买操作",
+        rows: 1,
+        id: "ruleContent_payAction",
+        type: "String",
+        hint: "填写JavaScript 返回购买链接或者调用购买接口",
+      },
+      {
+        title: "图片解密",
+        rows: 1,
+        id: "ruleContent_imageDecode",
+        type: "String",
+        hint: "填写JavaScript 返回解密图片的bytes ",
+      },
+    ],
+  },
+  review: {
+    name: "段评",
+    children: [
+      {
+        title: "段评URL",
+        rows: 1,
+        id: "ruleReview_reviewUrl",
+        type: "String",
+        hint: "段评URL",
+      },
+      {
+        title: "发布者头像",
+        rows: 1,
+        id: "ruleReview_avatarRule",
+        type: "String",
+        hint: "段评发布者头像",
+      },
+      {
+        title: "段评内容",
+        rows: 1,
+        id: "ruleReview_contentRule",
+        type: "String",
+        hint: "段评内容",
+      },
+      {
+        title: "发布时间",
+        rows: 1,
+        id: "ruleReview_postTimeRule",
+        type: "String",
+        hint: "段评发布时间",
+      },
+      {
+        title: "回复URL",
+        rows: 1,
+        id: "ruleReview_reviewQuoteUrl",
+        type: "String",
+        hint: "获取段评回复URL",
+      },
+      {
+        title: "点赞URL",
+        rows: 1,
+        id: "ruleReview_voteUpUrl",
+        type: "String",
+        hint: "点赞URL",
+      },
+      {
+        title: "点踩URL",
+        rows: 1,
+        id: "ruleReview_voteDownUrl",
+        type: "String",
+        hint: "点踩URL",
+      },
+      {
+        title: "发送回复",
+        rows: 1,
+        id: "ruleReview_postReviewUrl",
+        type: "String",
+        hint: "发送回复URL",
+      },
+      {
+        title: "回复段评",
+        rows: 1,
+        id: "ruleReview_postQuoteUrl",
+        type: "String",
+        hint: "发送回复段评URL",
+      },
+      {
+        title: "删除段评",
+        rows: 1,
+        id: "ruleReview_deleteUrl",
+        type: "String",
+        hint: "删除段评URL",
+      },
+    ],
+  },
+  other: {
+    name: "其他",
+    children: [
+      {
+        title: "启用搜索",
+        rows: 1,
+        id: "enabled",
+        type: "Boolean",
+        hint: "启用: true  关闭: false (可选,默认true)",
+      },
+      {
+        title: "启用发现",
+        rows: 1,
+        id: "enabledExplore",
+        type: "Boolean",
+        hint: "启用: true  关闭: false (可选,默认true)",
+      },
+      {
+        title: "启用段评",
+        rows: 1,
+        id: "enabledReview",
+        type: "Boolean",
+        hint: "启用: true  关闭: false (可选,默认false)",
+      },
+      {
+        title: "Cookie",
+        rows: 1,
+        id: "enabledCookieJar",
+        type: "Boolean",
+        hint: "保存返回头的Cookie 启用: true  关闭: false (可选,默认false)",
+      },
+      {
+        title: "搜索权重",
+        rows: 1,
+        id: "weight",
+        type: "Number",
+        hint: "整数: 0~N (可选,默认0) | 数字越大越靠前",
+      },
+      {
+        title: "排序编号",
+        rows: 1,
+        id: "customOrder",
+        type: "Number",
+        hint: "整数: 0~N (可选,默认0) | 数字越小越靠前",
+      },
+    ],
+  }
 };
